% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor.R
\name{chisqplot}
\alias{chisqplot}
\title{Takes in data and produces a chi-square plot by finding the mean squared distances and chi-square quantiles associated with the numeric variables.}
\usage{
chisqplot(data)
}
\arguments{
\item{...}{}
}
\value{
Uses numeric data variables to produce a chi-square plot in order to help assess normality.
}
\description{
Takes in data and produces a chi-square plot by finding the mean squared distances and chi-square quantiles associated with the numeric variables.
}
\examples{
chisqplot
}
