% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor.R
\name{dataSummary}
\alias{dataSummary}
\title{Basic Multivariate Data Statistics (number of qualitative and quantitave variable, numder of observations, mean and covariance etc. see vignette)}
\usage{
dataSummary(data)
}
\arguments{
\item{...}{}
}
\value{
List of statistics of the data
}
\description{
Basic Multivariate Data Statistics (number of qualitative and quantitave variable, numder of observations, mean and covariance etc. see vignette)
}
\examples{
dataSummary
}
