% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor.R
\name{myqq}
\alias{myqq}
\title{Q-Q plot for Numeric Data of four variables. Adds normal line to each plot in a 2x2 grid layout.(Used in shinyapp)}
\usage{
myqq(data)
}
\arguments{
\item{...}{}
}
\value{
Produces a Q-Qplot with normal line through the data on a 2x2 grid layout.
}
\description{
Q-Q plot for Numeric Data of four variables. Adds normal line to each plot in a 2x2 grid layout.(Used in shinyapp)
}
\examples{
myqq
}
