% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor.R
\name{myScatterPairs}
\alias{myScatterPairs}
\title{Produces scatterplots of selected pairs, with ellipse confidence intervals, of numeric variables (Used in shiny app)}
\usage{
myScatterPairs(data, x, y)
}
\arguments{
\item{...}{}
}
\value{
Produces a scatterplot of variables x and y from the data used.
}
\description{
Produces scatterplots of selected pairs, with ellipse confidence intervals, of numeric variables (Used in shiny app)
}
\examples{
myScatterPairs
}
