% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor.R
\name{myScatterPairsE}
\alias{myScatterPairsE}
\title{Produces scatterplots of selected pairs, with ellipse confidence intervals, of numeric variables (Used in shiny app)
Also provides a probability ellipse around the data with a specified alpha}
\usage{
myScatterPairsE(data, x, y, percent)
}
\arguments{
\item{...}{}
}
\value{
Produces a scatterplot of variables x and y from the data used. Uses percent to create a probability ellipse
}
\description{
Produces scatterplots of selected pairs, with ellipse confidence intervals, of numeric variables (Used in shiny app)
Also provides a probability ellipse around the data with a specified alpha
}
\examples{
myScatterPairsE 
}
