% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor.R
\name{OutlierRemovalE}
\alias{OutlierRemovalE}
\title{Removes Outliers from specified numeric data by creating a linear model and evaluating the cooks distance of the data (Used in shiny app)
Also provides a probability ellipse around the data with a specified alpha}
\usage{
OutlierRemovalE(data, x, y, percent)
}
\arguments{
\item{...}{}
}
\value{
Produces a scatterplot of variables x and y from the data used without detected outliers.  Uses percent to create a probability ellipse
}
\description{
Removes Outliers from specified numeric data by creating a linear model and evaluating the cooks distance of the data (Used in shiny app)
Also provides a probability ellipse around the data with a specified alpha
}
\examples{
OutlierRemovalE
}
